Fix negative duration in Activity
Fix overflow due to cast from int to short while extending the maximum
duration to the one in PatternActivity.

+++ mobitopp/src/test/java/edu/kit/ifv/mobitopp/simulation/activityschedule/ActivityTest.java
@@ -41,8 +41,8 @@ public class ActivityTest {
 												(byte)activityNr,
 												ActivityType.LEISURE,
 												startDate,
-												(short)duration,
-												(short)tripDuration,
+												duration,
+												tripDuration,
 												startFlexibility,
 												endFlexibility,
 												durationFlexibility
